Split source control filter hit testing
